updated docker

+++ blog-app-backend/app/Http/Controllers/BlogController.php
@@ -7,6 +7,7 @@ use App\Models\TempImage;
 use Illuminate\Support\Facades\File;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use CloudinaryLabs\CloudinaryLaravel\Facades\Cloudinary;
 
 class BlogController extends Controller
 {
@@ -24,6 +25,7 @@ class BlogController extends Controller
             'data' => $blogs
         ]);
     }
+
     // this method return a single blog
     public function show($id)
     {
@@ -40,7 +42,8 @@ class BlogController extends Controller
             'data' => $blog
         ]);
     }
-    //  this methos insert a  blog
+
+    // this method inserts a blog
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
@@ -55,6 +58,7 @@ class BlogController extends Controller
                 'errors' => $validator->errors()
             ]);
         }
+
         $blog = new Blog();
         $blog->title = $request->title;
         $blog->author = $request->author;
@@ -62,31 +66,32 @@ class BlogController extends Controller
         $blog->shortDesc = $request->shortDesc;
         $blog->save();
 
-        // save image here 
-        // $tempImage = TempImage::find($request->image_id);
-        // if ($tempImage != null) {
-        //     // create new image name
-        //     $imageExtArray = explode('.', $tempImage->name);
-        //     $ext = last($imageExtArray);
-        //     $imageName = time() . '-' . $blog->id . '.' . $ext;
-        //     // update db record
-        //     $blog->image = $imageName;
-        //     $blog->save();
-        //     // move file
-        //     $sourcePath = public_path('uploads/temp/' . $tempImage->name);
-        //     $destPath = public_path('uploads/blogs/' . $imageName);
-
-        //     if (File::copy($sourcePath, $destPath)) {
-        //         File::delete($sourcePath); // delete temp file
-        //         $tempImage->delete();      // optional: remove DB record of temp image
-        //     }
-        // }
-
+        // --- OLD Local Image Save Logic ---
+        /*
         $tempImage = TempImage::find($request->image_id);
         if ($tempImage != null) {
-            $blog->image = $tempImage->name; // already a Cloudinary URL
+            $imageExtArray = explode('.', $tempImage->name);
+            $ext = last($imageExtArray);
+            $imageName = time() . '-' . $blog->id . '.' . $ext;
+            $blog->image = $imageName;
             $blog->save();
-            $tempImage->delete(); // clean up
+
+            $sourcePath = public_path('uploads/temp/' . $tempImage->name);
+            $destPath = public_path('uploads/blogs/' . $imageName);
+
+            if (File::copy($sourcePath, $destPath)) {
+                File::delete($sourcePath);
+                $tempImage->delete();
+            }
+        }
+        */
+
+        // --- NEW Cloudinary Logic ---
+        $tempImage = TempImage::find($request->image_id);
+        if ($tempImage != null) {
+            $blog->image = $tempImage->name; // already Cloudinary URL
+            $blog->save();
+            $tempImage->delete();
         }
 
         return response()->json([
@@ -95,7 +100,8 @@ class BlogController extends Controller
             'data' => $blog
         ]);
     }
-    // this method update a blog
+
+    // this method updates a blog
     public function update($id, Request $request)
     {
         $blog = Blog::find($id);
@@ -118,67 +124,77 @@ class BlogController extends Controller
                 'errors' => $validator->errors()
             ]);
         }
+
         $blog->title = $request->title;
         $blog->author = $request->author;
         $blog->description = $request->description;
         $blog->shortDesc = $request->shortDesc;
         $blog->save();
 
-        // $tempImage = TempImage::find($request->image_id);
-        // if ($tempImage != null) {
-        //     // delete old image if exists
-        //     if ($blog->image && File::exists(public_path('uploads/blogs/' . $blog->image))) {
-        //         File::delete(public_path('uploads/blogs/' . $blog->image));
-        //     }
-        //     $imageExtArray = explode('.', $tempImage->name);
-        //     $ext = last($imageExtArray);
-        //     $imageName = time() . '-' . $blog->id . '.' . $ext;
-        //     $blog->image = $imageName;
-        //     $blog->save();
-
-        //     $sourcePath = public_path('uploads/temp/' . $tempImage->name);
-        //     $destPath = public_path('uploads/blogs/' . $imageName);
-
-        //     if (File::copy($sourcePath, $destPath)) {
-        //         File::delete($sourcePath); // delete temp file
-        //         $tempImage->delete();      // optional: remove DB record of temp image
-        //     }
-        // }
+        // --- OLD Local Update Logic ---
+        /*
         $tempImage = TempImage::find($request->image_id);
         if ($tempImage != null) {
-            $blog->image = $tempImage->name; // already a Cloudinary URL
+            if ($blog->image && File::exists(public_path('uploads/blogs/' . $blog->image))) {
+                File::delete(public_path('uploads/blogs/' . $blog->image));
+            }
+            $imageExtArray = explode('.', $tempImage->name);
+            $ext = last($imageExtArray);
+            $imageName = time() . '-' . $blog->id . '.' . $ext;
+            $blog->image = $imageName;
             $blog->save();
-            $tempImage->delete(); // clean up
+
+            $sourcePath = public_path('uploads/temp/' . $tempImage->name);
+            $destPath = public_path('uploads/blogs/' . $imageName);
+
+            if (File::copy($sourcePath, $destPath)) {
+                File::delete($sourcePath);
+                $tempImage->delete();
+            }
         }
+        */
+
+        // --- NEW Cloudinary Logic ---
+        $tempImage = TempImage::find($request->image_id);
+        if ($tempImage != null) {
+            $blog->image = $tempImage->name; // Cloudinary URL
+            $blog->save();
+            $tempImage->delete();
+        }
+
         return response()->json([
             'status' => true,
             'message' => 'blog updated successfully',
             'data' => $blog
         ]);
     }
-    // this method will delete a blog
+
+    // this method deletes a blog
     public function destroy($id)
     {
         $blog = Blog::find($id);
         if ($blog == null) {
-            return response()->json(
-                [
-                    'status' => false,
-                    'message' => 'Blog not found',
-                ]
-            );
+            return response()->json([
+                'status' => false,
+                'message' => 'Blog not found',
+            ]);
         }
-        // Delete Blog Image First
-        // File::delete(public_path('uploads/blogs/' . $blog->image));
 
-        // Delete Blog from DB
+        // --- OLD Local Delete Logic ---
+        /*
+        File::delete(public_path('uploads/blogs/' . $blog->image));
+        */
+
+        // --- NEW Cloudinary Delete (optional) ---
+        // If you want to delete from Cloudinary also:
+        // $publicId = pathinfo($blog->image, PATHINFO_FILENAME);
+        // Cloudinary::destroy('blogs/' . $publicId);
+
         $blog->delete();
 
-        return response()->json(
-            [
-                'status' => true,
-                'message' => 'Blog deleted  successfully'
-            ]
-        );
+        return response()->json([
+            'status' => true,
+            'message' => 'Blog deleted successfully'
+        ]);
     }
 }
